Fixed front screen issues
